+++ CHANGES.md
@@ -1,6 +1,6 @@
 ## v0.4.0
 
-* Add 4.07+rc1 information.
+* Add 4.07.0 release information.
 
 ## v0.3.0 (07/06/2018)